feat: separate delegate gig
